а

+++ Models/Country.cs
@@ -2,13 +2,11 @@
 
 namespace Travel_Company.WPF.Models;
 
-public partial class Country : ICatalogItem
+public partial class Country : ICatalogItem // Added implementation of ICatalogItem
 {
-    public int Id { get; set; }
-
+    public long Id { get; set; }
     public string Name { get; set; } = null!;
 
     public virtual ICollection<PopulatedPlace> PopulatedPlaces { get; set; } = new List<PopulatedPlace>();
-
     public virtual ICollection<Route> Routes { get; set; } = new List<Route>();
 }
